chore: organize structures
