+++ apps/admin-fnp/components/structures/columns/clients.tsx
@@ -6,7 +6,7 @@ import { PhoneNumberFormat } from "google-libphonenumber"
 import { ApplicationUser } from "@/lib/schemas"
 import { phoneUtility } from "@/lib/utilities"
 import { Checkbox } from "@/components/ui/checkbox"
-import { ControlDropDown } from "@/components/structures/control-dropdown"
+import { ControlDropDown } from "@/components/structures/dropdowns/control-dropdown"
 
 export const clientColumns: ColumnDef<ApplicationUser>[] = [
   {
