+++ apps/admin-fnp/components/structures/data-table.tsx
@@ -26,7 +26,7 @@ import {
 } from "@/components/ui/table"
 import { Icons } from "@/components/icons/lucide"
 
-import { Pagination } from "./pagination"
+import { Pagination } from "./controls/pagination"
 
 interface DataTableProps<TData, TValue> {
   columns: ColumnDef<TData, TValue>[]
@@ -133,9 +133,9 @@ export function DataTable<TData, TValue>({
                       {header.isPlaceholder
                         ? null
                         : flexRender(
-                            header.column.columnDef.header,
-                            header.getContext(),
-                          )}
+                          header.column.columnDef.header,
+                          header.getContext(),
+                        )}
                     </TableHead>
                   )
                 })}
